requirements.txt
python-telegram-bot
yt-dlp
fastapi
uvicorn
python-dotenv
slowapi
aiofiles
jinja2
